+++ calculator/src/main/java/com/pechuro/cashdebts/calculator/Calculator.kt
@@ -1,13 +1,11 @@
 package com.pechuro.cashdebts.calculator
 
-import java.math.BigDecimal
-
 interface Calculator {
 
     fun evaluate(expr: String): Result
 }
 
 sealed class Result {
-    class Success(val result: BigDecimal) : Result()
+    class Success(val result: Double) : Result()
     object Error : Result()
 }
